retinaai-backend/retinaai-backend/requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
pillow
pymysql
cryptography
reportlab
